fix: update urls in DB if site is being created from existing folder
